requirements.txt
django==2.0.7
djangorestframework==3.8.2
mysqlclient=1.3.13
django-cors-headers==2.3.0
